Created registers structure
