+++ src/virtual_machine/register.rs
@@ -1,5 +1,11 @@
 pub const AMOUNT_OF_REGISTERS: usize = 10;
 
+pub enum Flag {
+    POSITIVE = 1 << 0,
+    ZERO = 1 << 1,
+    NEGATIVE = 1 << 2,
+}
+
 #[derive(Clone, Copy)]
 pub enum Register {
     R0,
@@ -12,6 +18,36 @@ pub enum Register {
     R7,
     ProgramCounter,
     ConditionFlag,
+}
+
+pub struct Registers {
+    registers: Vec<u16>,
+}
+
+impl Registers {
+    pub fn new(program_counter_start: u16) -> Self {
+        let mut registers = vec![0; AMOUNT_OF_REGISTERS];
+        registers[Register::ProgramCounter as usize] = program_counter_start;
+        Self { registers }
+    }
+
+    pub fn read_register(&self, source_register: Register) -> u16 {
+        self.registers[source_register as usize]
+    }
+
+    pub fn update_register(&mut self, destination_register: Register, new_register_value: u16) {
+        self.registers[destination_register as usize] = new_register_value;
+    }
+
+    pub fn update_flags(&mut self, register: Register) {
+        if self.read_register(register) == 0 {
+            self.update_register(Register::ConditionFlag, Flag::ZERO as u16)
+        } else if (self.read_register(register) >> 15) != 0 {
+            self.update_register(Register::ConditionFlag, Flag::NEGATIVE as u16)
+        } else {
+            self.update_register(Register::ConditionFlag, Flag::POSITIVE as u16)
+        }
+    }
 }
 
 impl From<u16> for Register {
